% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{basesDifferent}
\alias{basesDifferent}
\title{Calculate the number of bases that are different between two GRanges objects.}
\usage{
basesDifferent(granges_1, granges_2)
}
\arguments{
\item{granges_1}{GRanges object}

\item{granges_2}{GRanges object}
}
\value{
data_frame with total_diff, and start_end_diff columns
}
\description{
Returns both total difference as well as the difference in start and end coordinates only.
}
\author{
Kaur Alasoo
}
