% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyDifference.R
\name{calculateBasepairDifference}
\alias{calculateBasepairDifference}
\title{Calculate the total number of nucleotides that are different between two transcripts.}
\usage{
calculateBasepairDifference(diff_object)
}
\arguments{
\item{diff_object}{Result object from the indentifyAddedRemovedRegions() function.}
}
\value{
Data frame of differences between two transcripts, stratified into upstream, contained and downstream parts.
}
\description{
Calculate the total number of nucleotides that are different between two transcripts.
}
