% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructAlternativeEvents.R
\name{constructAlternativeEventsWrapper}
\alias{constructAlternativeEventsWrapper}
\title{Apply constructAlternativeEvents() to an gene_id and metadata objects. Extend truncated transcripts first.}
\usage{
constructAlternativeEventsWrapper(gene_id, gene_metadata, exons, cdss,
  max_internal_diff = 10, max_start_end_diff = 25)
}
\arguments{
\item{gene_id}{Focal gene id.}

\item{gene_metadata}{Tibble containing gene metadata from the prepareAnnotations script.}

\item{exons}{List of annotated exons}

\item{cdss}{List of annotated CDSs}

\item{max_internal_diff}{maximal internal difference between two evets for them to be considered the same.}

\item{max_start_end_diff}{maximal difference at the start and end of two events for them to be considered 
the same.}
}
\value{
Three-level list. First level correspond the each set of overlapping transcripts, second 
level contains lists of downstream, upstream and contained alternative transcription events.
}
\description{
Apply constructAlternativeEvents() to an gene_id and metadata objects. Extend truncated transcripts first.
}
