% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendTruncatedTranscripts.R
\name{extendSingleTranscript}
\alias{extendSingleTranscript}
\title{Extend a single transcript in one direction using exons from the reference transcript}
\usage{
extendSingleTranscript(tx_id, longest_tx_id, direction, exons,
  max_exon_extension = 1e+05)
}
\arguments{
\item{tx_id}{ID of the truncated transcript.}

\item{longest_tx_id}{ID of the reference transcript.}

\item{direction}{Direction of extension (upstream or downstream)}

\item{exons}{GRangesList of transcripts.}

\item{max_exon_extension}{If truncated and reference transcripts exon bpundaries do not match,
then this specifies the maximum number of allowed nucleotides that the truncated transcript can be longer 
than the reference transcript for the extension to proceed.}
}
\value{
Extended version of the truncated transcript (GRanges object)
}
\description{
Extend a single transcript in one direction using exons from the reference transcript
}
