% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendTruncatedTranscripts.R
\name{extendTranscriptsPerGene}
\alias{extendTranscriptsPerGene}
\title{Extend truncated transcripts to the longest annotated transcript.}
\usage{
extendTranscriptsPerGene(metadata, exons, cdss)
}
\arguments{
\item{metadata}{data frame of transcript metadata (required columns: ensembl_transcript_id, 
longest_start, longest_end, cds_start_NF, cds_end_NF, cds_start_end_NF).}

\item{exons}{named list of GRanges objects containing the exon coordinates for each transcript.}

\item{cdss}{named list of GRanges objects containing the CDS coordinates for each transcript.}
}
\value{
List of GRanges lists (exons and cdss). Contains the update exon and CDS coordinates for 
transcripts that were previously truncated.
}
\description{
Identify transcripts that are flagged as start or end not found and extend them by adding 
missing exons from the longest transcript of the gene. Importantly, if the last exon
at the truncated end is not present in the longest transcript, then the truncated transcript 
will not be extended, because there is no information supporting that exon junction.
}
\author{
Kaur Alasoo
}

