% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{extractGeneData}
\alias{extractGeneData}
\title{Extract gene-specifc data from annotations data frame and exons/cdss lists.}
\usage{
extractGeneData(gene_id, annotations_df, exons, cdss, transcript_ids = NULL)
}
\arguments{
\item{gene_id}{Ensembl id of the gene.}

\item{annotations_df}{data frame containing transcript metadata.}

\item{exons}{named list of GRanges objects containing the exon coordinates for each transcript.}

\item{cdss}{named list of GRanges objects containing the CDS coordinates for each transcript.}
}
\value{
List of 3 objects: filtered gene metadata, filtered exon coordinates (gene trancripts only),
filtered CDS coordinates (gene transcripts only)
}
\description{
Extract gene-specifc data from annotations data frame and exons/cdss lists.
}
\author{
Kaur Alasoo
}
