% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructAlternativePromoters.R
\name{fillMissingInternalExons}
\alias{fillMissingInternalExons}
\title{Fill missing internal exons into alternative transcript start events}
\usage{
fillMissingInternalExons(transcripts)
}
\arguments{
\item{transcripts}{list of GRanges objects containing alternative transcript starts}
}
\value{
New list of alternative transcript starts where the missing exons have been filled in.
}
\description{
Fill missing internal exons into alternative transcript start events
}
\author{
Kaur Alasoo
}
