% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processEnsemblAnnotations.R
\name{filterTranscriptMetadata}
\alias{filterTranscriptMetadata}
\title{Prepare transcript metadata before extending truncated transcripts and constructing alternative events}
\usage{
filterTranscriptMetadata(transcript_metadata,
  complete_transcripts = c("protein_coding", "lincRNA"),
  incomplete_transcripts = c("nonsense_mediated_decay",
  "processed_transcript", "retained_intron"))
}
\arguments{
\item{transcript_metadata}{Data frame of transcript metadata.}

\item{complete_transcripts}{Biotypes of complete transcripts (CDS start and end NF flags extracted grom GFF)}

\item{incomplete_transcripts}{Biotypes of trnascripts whose start and end are likely to be missing.}
}
\value{
Filtered transcript metadata df.
}
\description{
Retain only genes with >= 2 transcripts and mark CDS start and end 'not found' for incomplete transcripts.
}
