% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructAlternativeEvents.R
\name{flattenAlternativeEvents}
\alias{flattenAlternativeEvents}
\title{Combine all alternative events into a single list with appropriate names}
\usage{
flattenAlternativeEvents(alt_events)
}
\description{
Combine all alternative events into a single list with appropriate names
}
