% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processEnsemblAnnotations.R
\name{importTranscriptMetadataFromGTF}
\alias{importTranscriptMetadataFromGTF}
\title{Extract transcript metadata from Ensembl GTF file}
\usage{
importTranscriptMetadataFromGTF(gtf_path, transcript_tags)
}
\arguments{
\item{gtf_path}{Path to the GTF file}

\item{transcript_tags}{tibble of transcript tags}
}
\value{
A tibble with transcript metadata.
}
\description{
Extract transcript metadata from Ensembl GTF file
}
