% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGRanges.R
\name{indentifyAddedRemovedRegions}
\alias{indentifyAddedRemovedRegions}
\title{Identify regions that are specific to one of the two transcripts.}
\usage{
indentifyAddedRemovedRegions(tx1_id, tx2_id, exons_list)
}
\arguments{
\item{tx1_id}{id of transcript 1.}

\item{tx2_id}{id of transcript 2.}

\item{exons_list}{named list of GRanges objects containing the exon coordinates for each transcript.}
}
\value{
List of GRanges object. First contains exons specific to transcript 1, second contains exons 
specifc to transcript 2, third contains the shared exons between the two transcripts.
}
\description{
Return NULL if there is no shared region between two transcripts.
}
\author{
Kaur Alasoo
}
