% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{listIntersect}
\alias{listIntersect}
\title{Calculate the intersection of a GRangesList object}
\usage{
listIntersect(granges_list)
}
\description{
Calculate the intersection of a GRangesList object
}
