% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTranscriptAnnotations.R
\name{makeGeneTranscriptFromExons}
\alias{makeGeneTranscriptFromExons}
\title{Construct gene or transcript annotations from exon annotations}
\usage{
makeGeneTranscriptFromExons(exons, parent_id, make_gene = FALSE)
}
\description{
Construct gene or transcript annotations from exon annotations
}

