% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processEnsemblAnnotations.R
\name{markLongestGencodeTranscript}
\alias{markLongestGencodeTranscript}
\title{For each gene in a gene metadata data frame, mark the transcript with longes sequence.}
\usage{
markLongestGencodeTranscript(gene_metadata)
}
\arguments{
\item{gene_metadata}{Data frame with original gene metadata (required columns:
ensembl_gene_id, ensembl_transcript_id, transcript_start, transcript_end, strand)}
}
\value{
Original metadata df with longest_start and longest_end columns added.
}
\description{
For each gene in a gene metadata data frame, mark the transcript with longes sequence.
}
