% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processEnsemblAnnotations.R
\name{markLongestTranscripts}
\alias{markLongestTranscripts}
\title{For each gene in a gene metadata data frame, mark the transcripts transcripts
that have the longest start and end positions.}
\usage{
markLongestTranscripts(gene_metadata)
}
\arguments{
\item{gene_metadata}{Data frame with original gene metadata (required columns:
ensembl_gene_id, ensembl_transcript_id, transcript_start, transcript_end, strand)}
}
\value{
Original metadata df with longest_start and longest_end columns added.
}
\description{
For each gene in a gene metadata data frame, mark the transcripts transcripts
that have the longest start and end positions.
}

