% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructAlternativeEvents.R
\name{mergeByMaxDifference}
\alias{mergeByMaxDifference}
\title{If two transcripts in a GRanges list a are not sufficiently different from each other, then keep only one.}
\usage{
mergeByMaxDifference(granges_list, max_internal_diff = 10,
  max_start_end_diff = 25)
}
\description{
If two transcripts in a GRanges list a are not sufficiently different from each other, then keep only one.
}
