% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{replaceExtendedTranscripts}
\alias{replaceExtendedTranscripts}
\title{Replace extended trancripts in gene-specific data list.}
\usage{
replaceExtendedTranscripts(gene_data, extended_transcripts)
}
\arguments{
\item{gene_data}{List produced by the extractGeneData function.}

\item{extended_transcripts}{List produced by the extendTranscriptsPerGene function.}
}
\value{
Update gene_data list where original transcripts have been replaced by extended transcripts.
}
\description{
Replace extended trancripts in gene-specific data list.
}
\author{
Kaur Alasoo
}
