% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTranscriptAnnotations.R
\name{transcriptsToAnnotations}
\alias{transcriptsToAnnotations}
\title{Convert list of exons to a transcript annotations that can be exported into a GFF file.}
\usage{
transcriptsToAnnotations(transcript_list, gene_tx_map)
}
\arguments{
\item{transcript_list}{Named list of GRanges objects containing exons of transcripts 
(transcripts ids are names of list elements)}

\item{gene_tx_map}{Data frame mapping transcript ids to gene ids (required columns: 
transcript_id, gene_id)}
}
\value{
Merged GRanges object containin proper metadata for GFF export.
}
\description{
Merge Granges object and construct gene and transcript entries. Add proper Parent-Child relationships.
}

